+++ fotoobo/tools/ems/monitor.py
@@ -7,13 +7,14 @@ from datetime import datetime
 from typing import Any, Dict
 
 from fotoobo.fortinet.forticlientems import FortiClientEMS
+from fotoobo.helpers.result import Result
 from fotoobo.helpers.config import config
 from fotoobo.inventory import Inventory
 
 log = logging.getLogger("fotoobo")
 
 
-def connections(host: str) -> Dict[Any, Any]:
+def connections(host: str) -> Result:
     """
     Get connections information from FortiClient EMS.
 
@@ -24,23 +25,27 @@ def connections(host: str) -> Dict[Any, Any]:
     Args:
         host (str): FortiClient EMS host defined in the inventory
 
-    Returns: Dict[str, Any]
+    Returns: Result
     """
+    result = Result()
     inventory = Inventory(config.inventory_file)
+
     ems: FortiClientEMS = inventory.get(host, "forticlientems")[host]
     ems.login()
-    response = ems.api("get", "/endpoints/connection/donut")
-    data: Dict[str, Any] = {}
-    data["data"] = list(response.json()["data"])
 
-    data["fotoobo"] = {}
+    response = ems.api("get", "/endpoints/connection/donut")
+
+    data: Dict[str, Any] = {"data": list(response.json()["data"]), "fotoobo": {}}
+
     for item in data["data"]:
         data["fotoobo"][item["token"]] = item["value"]
 
-    return data
+    result.push_result(host, data)
+
+    return result
 
 
-def endpoint_management_status(host: str) -> Dict[str, Any]:
+def endpoint_management_status(host: str) -> Result:
     """
     Get management information about endpoints registered in FortiClient EMS.
 
@@ -54,16 +59,17 @@ def endpoint_management_status(host: str) -> Dict[str, Any]:
     Args:
         host (str): FortiClient EMS host defined in the inventory
 
-    Returns: Dict[str, Any]
+    Returns: Result
     """
+    result = Result()
     inventory = Inventory(config.inventory_file)
+
     ems: FortiClientEMS = inventory.get(host, "forticlientems")[host]
     ems.login()
 
     response = ems.api("get", "/endpoints/management/donut")
 
-    data = {}
-    data["data"] = dict(response.json())["data"]
+    data = {"data": dict(response.json())["data"]}
 
     managed = unmanaged = 0
     for item in data["data"]:
@@ -76,10 +82,12 @@ def endpoint_management_status(host: str) -> Dict[str, Any]:
             log.debug("management: unmanaged: %s", unmanaged)
 
     data["fotoobo"] = {"managed": managed, "unmanaged": unmanaged}
-    return data
+
+    result.push_result(host, data)
+    return result
 
 
-def endpoint_online_outofsync(host: str) -> Dict[str, Any]:
+def endpoint_online_outofsync(host: str) -> Result:
     """
     Get amount of FortiClient EMS devices which are online but policy not in sync.
 
@@ -92,8 +100,9 @@ def endpoint_online_outofsync(host: str) -> Dict[str, Any]:
     Args:
         host (str): FortiClient EMS host defined in the inventory
 
-    Returns: Dict[str, Any]
+    Returns: Result
     """
+    result = Result()
     inventory = Inventory(config.inventory_file)
     ems: FortiClientEMS = inventory.get(host, "forticlientems")[host]
     ems.login()
@@ -105,10 +114,12 @@ def endpoint_online_outofsync(host: str) -> Dict[str, Any]:
     data = {"fotoobo": {"outofsync": response.json()["data"]["total"]}}
     log.debug("endpoints outofsync: %s", data["fotoobo"]["outofsync"])
 
-    return data
+    result.push_result(host, data)
+
+    return result
 
 
-def endpoint_os_versions(host: str) -> Dict[str, Any]:
+def endpoint_os_versions(host: str) -> Result:
     """
     Get management information about endpoints registered in FortiClient EMS.
 
@@ -125,13 +136,13 @@ def endpoint_os_versions(host: str) -> Dict[str, Any]:
 
     Returns: Dict[str, Any]
     """
+    result = Result()
     inventory = Inventory(config.inventory_file)
+
     ems: FortiClientEMS = inventory.get(host, "forticlientems")[host]
     ems.login()
 
-    data: Dict[str, Any] = {}
-    data["data"] = {}
-    data["fotoobo"] = {}
+    data: Dict[str, Any] = {"data": {}, "fotoobo": {}}
 
     for fctversion_os in ["fctversionwindows", "fctversionmac", "fctversionlinux"]:
         response = ems.api("get", f"/endpoints/{fctversion_os}/donut")
@@ -139,19 +150,23 @@ def endpoint_os_versions(host: str) -> Dict[str, Any]:
         count = sum(item["value"] for item in data["data"][fctversion_os])
         data["fotoobo"][fctversion_os] = count
 
-    return data
+    result.push_result(host, data)
+
+    return result
 
 
-def system(host: str) -> Dict[str, Any]:
+def system(host: str) -> Result:
     """
     Get system information from FortiClient EMS.
 
     Args:
         host (str): FortiClient EMS host defined in inventory
 
-    Returns: Dict[str, Any]
+    Returns: Result
     """
+    result = Result()
     inventory = Inventory(config.inventory_file)
+
     ems: FortiClientEMS = inventory.get(host, "forticlientems")[host]
     ems.login()
 
@@ -163,10 +178,12 @@ def system(host: str) -> Dict[str, Any]:
     response = ems.api("get", "/system/info/")
     log.debug("serial number: %s (from /system/info/)", response.json()["data"]["license"]["sn"])
 
-    return dict(response.json()["data"])
+    result.push_result(host, dict(response.json()["data"]))
+
+    return result
 
 
-def license(host: str) -> Dict[str, Any]:  # pylint: disable=redefined-builtin
+def license(host: str) -> Result:  # pylint: disable=redefined-builtin
     """
     Get license information from FortiClient EMS.
 
@@ -193,11 +210,13 @@ def license(host: str) -> Dict[str, Any]:  # pylint: disable=redefined-builtin
     Args:
         host (str): FortiClient EMS host defined in the inventory
 
-    Returns: Dict[str, Any]
+    Returns: Result
     """
+    result = Result()
     inventory = Inventory(config.inventory_file)
     ems: FortiClientEMS = inventory.get(host, "forticlientems")[host]
     ems.login()
+
     response = ems.api("get", "/license/get")
     data = dict(response.json()["data"])
 
@@ -211,6 +230,7 @@ def license(host: str) -> Dict[str, Any]:  # pylint: disable=redefined-builtin
 
     data["fotoobo"] = {}
     data["fotoobo"]["license_expiry_days"] = license_expiry_days
+
     for key in data["seats"]:
         if data["seats"][key] > 0:
             log.debug(f"{key} license count: %s", data["seats"][key])
@@ -219,4 +239,6 @@ def license(host: str) -> Dict[str, Any]:  # pylint: disable=redefined-builtin
             log.debug(f"{key} license usage : %s%%", license_usage)
             data["fotoobo"][key + "_usage"] = license_usage
 
-    return data
+    result.push_result(host, data)
+
+    return result