+++ tests/cli/fgt/test_cli_fgt_config.py
@@ -1,5 +1,5 @@
 """
-Testing the cli fgt config check
+Test the cli fgt config
 """